pyproject.toml
[tool.poetry]
name = "pokeql"
version = "0.1.0"
description = ""
authors = ["Jaime Abbariao <jaimeabbariao@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
strawberry-graphql = "^0.93.10"
websockets = "^10.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
